requirements.txt

torch
torchvision
Pillow
